Feat/ios 3 (#3)
* DetailsView

* Views Structure

* Auth

* Virtual Trainer

* chart view

* api error

* api error handling

+++ code/backend/src/application/qb8s.net.OptiFit.CQRS/Dtos/Workout/GetWorkoutDto.cs
@@ -23,7 +23,7 @@ public class GetWorkoutDtoProfile : BaseProfile
         CreateMap<Core.Entities.Workout, WorkoutSummary>()
             .ForMember(dest => dest.TotalTime,
                 opt => opt.MapFrom(src => src.EndAtUtc.HasValue
-                    ? (int)(src.EndAtUtc.Value - src.StartAtUtc).TotalMinutes
+                    ? (src.EndAtUtc.Value - src.StartAtUtc).TotalMinutes
                     : 0))
             .ForMember(dest => dest.TotalExercises,
                 opt => opt.MapFrom(src => src.WorkoutExercises.Count))
