+++ code/backend/src/application/qb8s.net.OptiFit.CQRS/Dtos/Workout/WorkoutSummary.cs
@@ -2,7 +2,7 @@ namespace qb8s.net.OptiFit.CQRS.Dtos.Workout;
 
 public class WorkoutSummary
 {
-    public int TotalTime { get; set; }
+    public double TotalTime { get; set; }
     public int TotalSets { get; set; }
     public int TotalReps { get; set; }
     public decimal TotalWeight { get; set; }
